+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file, following t
 ## [Unreleased]
 ### Added
 - `backgroundColor` parameter for `Stage`
+- x/y/z offset parameters for labels (TextBuffer)
 - `OX1` atoms are recognized as part of the protein backbone
 - `stage.makeImage` now increments the task counter
 - added `.isIdentity` method to test if an `Assembly` is an identity transformation over all chains